refactor(): move timer badge datasource into timer badge collection view
- Remove datasource in all controller using timer badge collection view
