+++ timer/Source/Module/TimeSetDetail/TimeSetDetailViewController.swift
@@ -36,19 +36,6 @@ class TimeSetDetailViewController: BaseHeaderViewController, View {
     // MARK: - properties
     var coordinator: TimeSetDetailViewCoordinator
     
-    private lazy var dataSource = RxCollectionViewSectionedAnimatedDataSource<TimerBadgeSectionModel>(configureCell: { (dataSource, collectionView, indexPath, cellType) -> UICollectionViewCell in
-        switch cellType {
-        case let .regular(reactor):
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TimerBadgeCollectionViewCell.name, for: indexPath) as? TimerBadgeCollectionViewCell else { fatalError() }
-            cell.reactor = reactor
-            
-            return cell
-            
-        case .extra(_):
-            fatalError("This view can't present extra cells of timer badge collection view")
-        }
-    })
-    
     // MARK: - constructor
     init(coordinator: TimeSetDetailViewCoordinator) {
         self.coordinator = coordinator
@@ -155,7 +142,7 @@ class TimeSetDetailViewController: BaseHeaderViewController, View {
         reactor.state
             .filter { $0.shouldSectionReload }
             .map { $0.sections }
-            .bind(to: timerBadgeCollectionView.rx.items(dataSource: dataSource))
+            .bind(to: timerBadgeCollectionView.rx.items(dataSource: timerBadgeCollectionView._dataSource))
             .disposed(by: disposeBag)
         
         let selectedIndex = reactor.state
