+++ timer/Source/Module/TimeSetProcess/TimeSetProcessViewController.swift
@@ -53,18 +53,6 @@ class TimeSetProcessViewController: BaseHeaderViewController, View {
     // MARK: - properties
     var coordinator: TimeSetProcessViewCoordinator
     
-    private lazy var dataSource = RxCollectionViewSectionedAnimatedDataSource<TimerBadgeSectionModel>(configureCell: { (dataSource, collectionView, indexPath, cellType) -> UICollectionViewCell in
-        switch cellType {
-        case let .regular(reactor):
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TimerBadgeCollectionViewCell.name, for: indexPath) as? TimerBadgeCollectionViewCell else { fatalError() }
-            cell.reactor = reactor
-            return cell
-            
-        case .extra(_):
-            fatalError("This view can't present extra cells of timer badge collection view")
-        }
-    })
-    
     // Dispose bags
     private var popupDisposeBag = DisposeBag()
     private var alertDisposeBag = DisposeBag()
@@ -254,7 +242,7 @@ class TimeSetProcessViewController: BaseHeaderViewController, View {
         reactor.state
             .filter { $0.shouldSectionReload }
             .map { $0.sections }
-            .bind(to: timerBadgeCollectionView.rx.items(dataSource: dataSource))
+            .bind(to: timerBadgeCollectionView.rx.items(dataSource: timerBadgeCollectionView._dataSource))
             .disposed(by: disposeBag)
         
         reactor.state
