+++ timer/Source/Module/TimeSetSave/TimeSetSaveViewController.swift
@@ -34,19 +34,6 @@ class TimeSetSaveViewController: BaseHeaderViewController, View {
     // MARK: - properties
     var coordinator: TimeSetSaveViewCoordinator
     
-    private lazy var dataSource = RxCollectionViewSectionedAnimatedDataSource<TimerBadgeSectionModel>(configureCell: { (dataSource, collectionView, indexPath, cellType) -> UICollectionViewCell in
-        switch cellType {
-        case let .regular(reactor):
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TimerBadgeCollectionViewCell.name, for: indexPath) as? TimerBadgeCollectionViewCell else { fatalError() }
-            cell.reactor = reactor
-            
-            return cell
-            
-        case .extra(_):
-            fatalError("This view can't present extra cells of timer badge collection view")
-        }
-    })
-    
     // MARK: - constructor
     init(coordinator: TimeSetSaveViewCoordinator) {
         self.coordinator = coordinator
@@ -165,7 +152,7 @@ class TimeSetSaveViewController: BaseHeaderViewController, View {
         reactor.state
             .filter { $0.shouldSectionReload }
             .map { $0.sections }
-            .bind(to: timerBadgeCollectionView.rx.items(dataSource: dataSource))
+            .bind(to: timerBadgeCollectionView.rx.items(dataSource: timerBadgeCollectionView._dataSource))
             .disposed(by: disposeBag)
         
         reactor.state
